requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.4
djangorestframework==3.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.7
queuelib==1.5.0
requests==2.21.0
Scrapy==1.5.1
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.19.0
zope.interface==4.6.0
